Implement update functionality for product category and description
This commit includes new functionality to update both the product category and the category description. This was facilitated by addition of new endpoints in the ProductController and CategoryController classes. Corresponding service methods were also created. Credit to these changes, users are now able to modify product category and their descriptions improving the flexibility of the application.

Implement update functionality for product category and description

This commit includes new functionality to update both the product category and the category description. This was facilitated by addition of new endpoints in the ProductController and CategoryController classes. Corresponding service methods were also created. Credit to these changes, users are now able to modify product category and their descriptions improving the flexibility of the application.